refactor: rename package in :feature:licenses
