+++ composeApp/src/androidMain/kotlin/com/livefast/eattrash/raccoonforfriendica/di/DiHelper.kt
@@ -49,7 +49,7 @@ import com.livefast.eattrash.raccoonforfriendica.feature.timeline.di.featureTime
 import com.livefast.eattrash.raccoonforfriendica.feature.unpublished.di.featureUnpublishedModule
 import com.livefast.eattrash.raccoonforfriendica.feature.userlist.di.featureUserListModule
 import com.livefast.eattrash.raccoonforfriendica.feaure.search.di.featureSearchModule
-import com.livefast.eattrash.raccoonforlemmy.unit.licences.di.featureLicenceModule
+import com.livefast.eattrash.raccoonforfriendica.unit.licences.di.featureLicenceModule
 import org.koin.dsl.module
 
 val sharedHelperModule =
@@ -64,7 +64,6 @@ val sharedHelperModule =
             coreL10nModule,
             corePersistenceModule,
             corePreferencesModule,
-            domainPushNotificationsModule,
             coreNavigationModule,
             coreNotificationsModule,
             coreResourceModule,
