+++ feature/licences/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/unit/licences/LicenceUrls.kt
@@ -1,4 +1,4 @@
-package com.livefast.eattrash.raccoonforlemmy.unit.licences
+package com.livefast.eattrash.raccoonforfriendica.unit.licences
 
 internal object LicenceUrls {
     const val ANDROIDX = "https://github.com/androidx/androidx/blob/androidx-main/LICENSE.txt"
