+++ feature/licences/src/commonMain/kotlin/com/livefast/eattrash/raccoonforfriendica/unit/licences/LicencesMviModel.kt
@@ -1,8 +1,8 @@
-package com.livefast.eattrash.raccoonforlemmy.unit.licences
+package com.livefast.eattrash.raccoonforfriendica.unit.licences
 
 import cafe.adriel.voyager.core.model.ScreenModel
 import com.livefast.eattrash.raccoonforfriendica.core.architecture.MviModel
-import com.livefast.eattrash.raccoonforlemmy.unit.licences.models.LicenceItem
+import com.livefast.eattrash.raccoonforfriendica.unit.licences.models.LicenceItem
 
 interface LicencesMviModel :
     ScreenModel,
